Refactor Profile component and update styles
Moved Profile component and its styles to a dedicated Profile directory for better organization. Updated imports in App.tsx to reflect new paths. Adjusted light theme colors in index.css and removed Profile-related styles from global CSS, placing them in Profile/Profile.css.
