+++ lab2/src/components/Image.tsx
@@ -1,14 +1,14 @@
-import React from 'react';
+import React from "react";
 
 type ImageProps = {
   src: string;
   alt: string;
   style?: React.CSSProperties;
   className?: string;
-}
+};
 
-const Image = ({src,alt, ...props }:ImageProps) => {
-  return <img src={src} alt={alt} {...props} />
-}
+const Image = ({ src, alt, ...props }: ImageProps) => {
+  return <img src={src} alt={alt} {...props} />;
+};
 
 export default Image;
