+++ lab2/src/index.css
@@ -49,8 +49,8 @@ button:focus-visible {
 
 @media (prefers-color-scheme: light) {
   :root {
-    color: #213547;
-    background-color: #ffffff;
+    color: #e7e7e7;
+    background-color: #666688;
   }
   a:hover {
     color: #747bff;
@@ -58,21 +58,4 @@ button:focus-visible {
   button {
     background-color: #f9f9f9;
   }
-}
-
-.img{
-    border-radius: 50%;
-    max-width: 100px;
-    max-height: 100px;
-}
-
-.profile{
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    border: 1px solid #ccc;
-    padding: 20px;
-    border-radius: 10px;
-    width: 250px;
-    background-color: #1f232c;
 }
